implement git-style paging for list output
Implements paging (a la 'less' or 'more') for the output of the list
commands.  The implementation is based on the implementation of
paging in the git cli.  The flag --no-pager is added to prevent
any paging, and the pager command defaults to less -FRX, unless
specified by the $PAGER env variable (both match the git implementation).

Signed-off-by: Daniel Harms <jdharms@gmail.com>