feat: add support for common v2 endpoints (ping/metrics/version/config)
This commit adds v2 (Ireland/Jakarta) support for four common endpoints:
- version
- metrics
- ping
- config

For information about the available commands, use `edgex-cli -h`

There are a number of ways to use the commands:
1. `edgex-cli version` displays version of each active microservice.
The version commands just shows the version string. The other commands
use a pretty-printed JSON structure.

2. `edgex-cli version -d` shows debug output - the URL used and the result

3. `edgex-cli version -j` gets the raw JSON output for one service (by default core-data)

It's also possible to specify the service to use. : `-m/--metadata`, `-c/--command`,
 `-n/--notification`, `-s/--scheduler` or `--data` (which is the default).
This can also be combined with `-j` or `-d`

Example:
- `$ edgex-cli metrics -j -m`
- `$ edgex-cli ping -d --command`

Help is available with
```
$ edgex-cli help
$ edgex-cli version --help
$ edgex-cli version --h
```

Signed-off-by: Siggi Skulason <siggi.skulason@canonical.com>