feat: add support for core-command read
This commit adds support for the `read` command of the `command` endpoint.

```
edgex-cli command read
```

Adds support for the `-j`, `-v` flags, which generate raw JSON output.
Adds support for the `-p` flag; if set, a successful GET will result in
an event being pushed to the EdgeX system
Adds support for the `-r` flag; if set, there will be no Event returned
in the http response

Signed-off-by: Mengyi <mengyi.wang@canonical.com>