feat: add support for core-command list
This commit adds support for the `list` command of the `command` endpoint.

```
edgex-cli command list
```

Adds support for the `-j`, `-v` flags, which generate raw JSON output.
Adds support for the `-l` flag, which limits the number of items to return.
Adds support for the `-o` flag, which indicates the number of items to skip.

Signed-off-by: Mengyi <mengyi.wang@canonical.com>